+++ src/components/common/index.ts
@@ -1,2 +1,3 @@
+export { default as Row } from './Row';
 export { default as SubTitle } from './SubTitle';
 export { default as Title } from './Titlte';
